requirements.txt
flask
flask-limiter
flask-wtf
psycopg2-binary
python-dotenv
pandas
openpyxl
